Add showDetails

+++ kodilla-collections/src/main/java/com/kodilla/collections/interfaces/homework/Car.java
@@ -1,7 +1,7 @@
 package com.kodilla.collections.interfaces.homework;
 
 public interface Car {
-     double getSpeed();
+     int getSpeed();
      int increaseSpeed();
      int decreaseSpeed();
 }
